+++ Core/Systems/Conversions/ImageConversionSystem.cs
@@ -1,3 +1,5 @@
+using System;
+using UGUIDots.Collections.Runtime;
 using UGUIDots.Render;
 using Unity.Mathematics;
 using UnityEngine;
@@ -12,27 +14,38 @@ namespace UGUIDots.Conversions.Systems {
     public class ImageConversionSystem : GameObjectConversionSystem {
 
         protected override void OnUpdate() {
-            Entities.ForEach((Image image) => {
-                TextureBin.TryLoadTextureBin("TextureBin", out TextureBin textureBin);
+            // TODO: Find a way to handle loading bins that hold textures and materials instead of using
+            // Resources.Load
+            var textureLoad  = TextureBin.TryLoadBin("TextureBin", out Bin<Texture> textureBin);
+            var materialLoad = MaterialBin.TryLoadBin("MaterialBin", out var materialBin);
 
-                var texture    = image.sprite != null ? image.sprite.texture : Texture2D.whiteTexture;
-                var imageIndex = textureBin.Add(texture);
+#if UNITY_EDITOR
+            if (!textureLoad) {
+                throw new InvalidOperationException("TextureBin is not located in: Assets/Resources");
+            }
+            if (!materialLoad) {
+                throw new InvalidOperationException("MaterialBin is not located in: Assets/Resources");
+            }
+#endif
+
+            Entities.ForEach((Image image) => {
+
+                var texture  = image.sprite != null ? image.sprite.texture : (Texture)Texture2D.whiteTexture;
+                var imageKey = (short)textureBin.Add(texture);
+
+                var material    = image.material != null ? image.material : Canvas.GetDefaultCanvasMaterial();
+                var materialKey = (short)materialBin.Add(material);
 
                 var entity   = GetPrimaryEntity(image);
-                var material = image.material != null ? image.material : Canvas.GetDefaultCanvasMaterial();
-                DstEntityManager.AddComponentObject(entity, material);
-
-                // TODO: Internally this would need a look up table...
-                // DstEntityManager.AddSharedComponentData(entity, new MaterialID { Value = material.GetInstanceID() });
-
                 var rectSize = image.rectTransform.Int2Size();
 
-                DstEntityManager.AddComponentData(entity, new TextureKey   { Value = imageIndex });
+                DstEntityManager.AddComponentData(entity, new TextureKey   { Value = imageKey });
+                DstEntityManager.AddComponentData(entity, new MaterialKey  { Value = materialKey });
                 DstEntityManager.AddComponentData(entity, new AppliedColor { Value = image.color });
                 DstEntityManager.AddComponentData(entity, new Dimensions   { Value = rectSize });
 
                 var spriteTexture = image.sprite;
-                var spriteRes = spriteTexture != null ? 
+                var spriteRes = spriteTexture != null ?
                     new int2(spriteTexture.texture.width, spriteTexture.texture.height) :
                     rectSize;
 
@@ -42,8 +55,8 @@ namespace UGUIDots.Conversions.Systems {
                 DstEntityManager.AddComponentData(entity, spriteData);
 
                 // TODO: Does not handle image slicing
-                DstEntityManager.AddBuffer<MeshVertexData>(entity).ResizeUninitialized(4);
-                DstEntityManager.AddBuffer<TriangleIndexElement>(entity).ResizeUninitialized(6);
+                DstEntityManager.AddBuffer<LocalVertexData>(entity).ResizeUninitialized(4);
+                DstEntityManager.AddBuffer<LocalTriangleIndexElement>(entity).ResizeUninitialized(6);
             });
         }
     }