+++ Core/Systems/Conversions/RectTransformConversionSystem.cs
@@ -1,5 +1,4 @@
 using UGUIDots.Transforms;
-using Unity.Transforms;
 using UnityEngine;
 
 namespace UGUIDots.Conversions.Systems {